+++ packages/expo-localization/ios/LocalizationModule.swift
@@ -142,7 +142,8 @@ public class LocalizationModule: Module {
           "digitGroupingSeparator": userSettingsLocale.groupingSeparator,
           "measurementSystem": getMeasurementSystemForLocale(userSettingsLocale),
           "currencyCode": languageLocale.currencyCode,
-          "currencySymbol": languageLocale.currencySymbol
+          "currencySymbol": languageLocale.currencySymbol,
+          "temperatureUnit": getTemperatureUnit()
         ]
       }
   }
@@ -152,6 +153,20 @@ public class LocalizationModule: Module {
     // we send both events since on iOS it means both calendar and locale needs an update
     sendEvent(LOCALE_SETTINGS_CHANGED)
     sendEvent(CALENDAR_SETTINGS_CHANGED)
+  }
+
+  static func getTemperatureUnit() -> String? {
+    let formatter = MeasurementFormatter()
+    formatter.locale = Locale.current
+
+    let temperature = Measurement(value: 0, unit: UnitTemperature.celsius)
+    let formatted = formatter.string(from: temperature)
+
+    guard let unitCharacter = formatted.last else {
+      return nil
+    }
+
+    return unitCharacter == "F" ? "fahrenheit" : "celsius"
   }
 
   // https://stackoverflow.com/a/28183182